Extract some common logic to a common file.
